Refactor editor props validation
